Move achievement methods into AchievementModel; add fields for rounds, riichi stat etc.
