+++ Mimir/src/models/Event.php
@@ -84,475 +84,6 @@ class EventModel extends Model
     }
 
     /**
-     * Get achievements list
-     * @throws AuthFailedException
-     * @throws \Exception
-     * @param $eventIdList
-     * @return array
-     */
-    public function getAchievements($eventIdList)
-    {
-        if (!$this->checkAdminToken()) {
-            throw new AuthFailedException('Only administrators are allowed to view achievements');
-        }
-
-        return [
-            'bestHand' => AchievementsPrimitive::getBestHandOfEvent($this->_db, $eventIdList),
-            'bestTsumoist' => AchievementsPrimitive::getBestTsumoistInSingleSession($this->_db, $eventIdList),
-            'braveSapper' => AchievementsPrimitive::getBraveSappers($this->_db, $eventIdList),
-            'dieHard' => AchievementsPrimitive::getDieHardData($this->_db, $eventIdList),
-            'dovakin' => AchievementsPrimitive::getDovakins($this->_db, $eventIdList),
-            'yakuman' => AchievementsPrimitive::getYakumans($this->_db, $eventIdList),
-            'shithander' => AchievementsPrimitive::getBestShithander($this->_db, $eventIdList),
-            'bestDealer' => AchievementsPrimitive::getBestDealer($this->_db, $eventIdList),
-            'bestFu' => AchievementsPrimitive::getMaxFuHand($this->_db, $eventIdList),
-            'impossibleWait' => AchievementsPrimitive::getImpossibleWait($this->_db, $eventIdList),
-            'honoredDonor' => $this->getHonoredDonor($eventIdList),
-            'justAsPlanned' => AchievementsPrimitive::getJustAsPlanned($this->_db, $eventIdList),
-            'carefulPlanning' => $this->getMinFeedsScore($eventIdList),
-            'doraLord' => AchievementsPrimitive::getMaxAverageDoraCount($this->_db, $eventIdList),
-            'catchEmAll' => AchievementsPrimitive::getMaxDifferentYakuCount($this->_db, $eventIdList),
-            'favoriteAsapinApprentice' => AchievementsPrimitive::getFavoriteAsapinApprentice($this->_db, $eventIdList),
-            'andYourRiichiBet' => $this->getMaxStolenRiichiBetsCount($eventIdList),
-            'prudent' => $this->getMinLostRiichiBetsPercentage($eventIdList)
-        ];
-    }
-
-    /**
-     * Get players who lost largest number of points as riichi bets
-     *
-     * @param $eventIdList
-     * @return array
-     * @throws \Exception
-     */
-    protected function getHonoredDonor($eventIdList)
-    {
-        $riichiStat = $this->_getRiichiStat($eventIdList);
-
-        uasort($riichiStat, function ($a, $b) {
-            return $a['lost'] < $b['lost'] ? 1 : -1;
-        });
-
-        return array_map(
-            function ($item) {
-                return [
-                    'name'  => $item['name'],
-                    'count'  => $item['lost'],
-                ];
-            },
-            array_slice($riichiStat, 0, 5)
-        );
-    }
-
-    /**
-     * Get players who has smallest score (average number) of feeding into others' hands
-     *
-     * @param $eventIdList
-     * @return array
-     * @throws \Exception
-     */
-    protected function getMinFeedsScore($eventIdList)
-    {
-        $games = SessionPrimitive::findByEventListAndStatus(
-            $this->_db,
-            $eventIdList,
-            SessionPrimitive::STATUS_FINISHED
-        );
-
-        $sessionIds = array_map(function (SessionPrimitive $el) {
-            return $el->getId();
-        }, $games);
-
-        $rounds = $this->_getRounds($sessionIds); // key: session id, value: RoundPrimitive[]
-        $payments = [];
-        foreach ($games as $session) {
-            $lastRound = null;
-            foreach ($rounds[$session->getId()] as $round) {
-                if ($lastRound === null) {
-                    $lastRound = $round;
-                    continue;
-                }
-
-                $lastSessionState = $round->getLastSessionState();
-                $payments = $this->_addLoserPayment($lastRound, $lastSessionState, $payments);
-                $lastRound = $round;
-            }
-
-            $payments = $this->_addLoserPayment($lastRound, $session->getCurrentState(), $payments);
-        }
-
-        $feedsScores = [];
-        foreach ($payments as $key => $value) {
-            $feedsScores[$key] = $value['sum'] / $value['count'];
-        }
-
-        asort($feedsScores);
-        $players = EventModel::getPlayersOfGames($this->_db, $games);
-
-        return array_map(
-            function ($playerId, $feedsScore) use ($players) {
-                return [
-                    'name'  => $players[$playerId]['display_name'],
-                    'score' => round($feedsScore)
-                ];
-            },
-            array_slice(array_keys($feedsScores), 0, 5),
-            array_slice(array_values($feedsScores), 0, 5)
-        );
-    }
-
-    /**
-     * Get players who collected the largest amount of other players' riichi bets during the tournament
-     *
-     * @param $eventIdList
-     * @return array
-     * @throws \Exception
-     */
-    protected function getMaxStolenRiichiBetsCount($eventIdList)
-    {
-        $riichiStat = $this->_getRiichiStat($eventIdList);
-
-        $filteredRiichiStat = array_filter($riichiStat, function ($item) {
-            return $item['stole'] > 0;
-        });
-
-        uasort($filteredRiichiStat, function ($a, $b) {
-            return $a['stole'] < $b['stole'] ? 1 : -1;
-        });
-
-        return array_map(
-            function ($item) {
-                return [
-                    'name'  => $item['name'],
-                    'count' => $item['stole'],
-                ];
-            },
-            array_slice($filteredRiichiStat, 0, 5)
-        );
-    }
-
-    /**
-     * Get players with losing the smallest percentage of riichi bets
-     *
-     * @param $eventIdList
-     * @return array
-     * @throws \Exception
-     */
-    protected function getMinLostRiichiBetsPercentage($eventIdList)
-    {
-        $riichiStat = $this->_getRiichiStat($eventIdList);
-
-        foreach ($riichiStat as &$item) {
-            $riichiBetsCount = $item['won'] +  $item['lost'];
-
-            $score = $riichiBetsCount !== 0
-                ? $item['lost'] / $riichiBetsCount * 100
-                : -1;
-
-            $item['score'] = $score;
-            $item['total'] = $riichiBetsCount;
-        }
-
-        uasort($riichiStat, function ($a, $b) {
-            return $a['score'] > $b['score'] ? 1 : -1;
-        });
-
-        $lostPercentages = array_filter($riichiStat, function ($item) {
-            return $item['score'] !== -1;
-        });
-
-        return array_map(
-            function ($item) {
-                return [
-                    'name'  => $item['name'],
-                    'score' => sprintf('%.2f', $item['score']),
-                    'lost'  => $item['lost'],
-                    'total' => $item['total'],
-                ];
-            },
-            array_slice($lostPercentages, 0, 5)
-        );
-    }
-
-    /**
-     * @param $eventIdList
-     * @return array
-     * @throws \Exception
-     */
-    protected function _getRiichiStat($eventIdList)
-    {
-        $games = SessionPrimitive::findByEventListAndStatus(
-            $this->_db,
-            $eventIdList,
-            SessionPrimitive::STATUS_FINISHED
-        );
-
-        $sessionIds = array_map(function (SessionPrimitive $el) {
-            return $el->getId();
-        }, $games);
-
-        $rounds = $this->_getRounds($sessionIds); // key: session id, value: RoundPrimitive[]
-        $players = EventModel::getPlayersOfGames($this->_db, $games);
-
-        $riichiStat = [];
-        foreach ($players as $playerId => $player) {
-            $riichiStat[$playerId] = [
-                'name'  => $player['display_name'],
-                'won'   => 0,
-                'lost'  => 0,
-                'stole' => 0
-            ];
-        }
-
-        foreach ($games as $session) {
-            $rules = $session->getEvent()->getRuleset();
-            if ($rules->riichiGoesToWinner()) {
-                $firstPlace = array_filter($session->getSessionResults(), function (SessionResultsPrimitive $result) {
-                    return $result->getPlace() === 1;
-                });
-
-                $firstPlacePlayerId = array_values($firstPlace)[0]->getPlayerId();
-                $riichiStat[$firstPlacePlayerId]['stole'] += $session->getCurrentState()->getRiichiBets();
-            }
-
-            foreach ($rounds[$session->getId()] as $round) {
-                $riichiIds = $round->getRiichiIds();
-
-                if (in_array($round->getOutcome(), ['ron', 'tsumo'])) {
-                    $winnerId = $round->getWinnerId();
-
-                    foreach ($riichiIds as $riichiPlayerId) {
-                        if ($riichiPlayerId == $winnerId) {
-                            $riichiStat[$riichiPlayerId]['won'] ++;
-                        } else {
-                            $riichiStat[$riichiPlayerId]['lost'] ++;
-                        }
-                    }
-                }
-
-                if (in_array($round->getOutcome(), ['draw', 'abort'])) {
-                    foreach ($riichiIds as $riichiPlayerId) {
-                        $riichiStat[$riichiPlayerId]['lost'] ++;
-                    }
-                }
-
-                if ($round->getOutcome() == 'multiron') {
-                    /** @var $round MultiRoundPrimitive */
-                    $riichiIds = $round->getRiichiIds();
-                    $winnerIds = $round->getWinnerIds();
-
-                    foreach ($riichiIds as $playerId) {
-                        if (!in_array($playerId, $winnerIds)) {
-                            $riichiStat[$playerId]['lost'] ++;
-                        }
-                    }
-
-                    $lastSessionState = $round->getLastSessionState();
-                    $riichiWinners = PointsCalc::assignRiichiBets(
-                        $round->rounds(),
-                        $round->getLoserId(),
-                        $lastSessionState->getRiichiBets(),
-                        $lastSessionState->getHonba(),
-                        $round->getSession()
-                    );
-
-                    foreach ($riichiWinners as $winnerId => $item) {
-                        $closestWinner = $item['closest_winner'];
-
-                        if ($rules->doubleronRiichiAtamahane() && $closestWinner) {
-                            if ($closestWinner == $winnerId) {
-                                if (in_array($winnerId, $riichiIds)) {
-                                    $riichiStat[$winnerId]['won'] ++;
-                                }
-
-                                $fromOthers = array_filter($riichiIds, function ($playerId) use ($winnerId) {
-                                    return $playerId != $winnerId;
-                                });
-
-                                $riichiStat[$winnerId]['stole'] += count($fromOthers);
-                            } else {
-                                $riichiStat[$winnerId]['lost'] ++;
-                            }
-                        } else {
-                            if (in_array($winnerId, $riichiIds)) {
-                                $riichiStat[$winnerId]['won'] ++;
-                            }
-
-                            $fromOthers = array_filter($item['from_players'], function ($playerId) use ($winnerId) {
-                                return $playerId != $winnerId;
-                            });
-
-                            $riichiStat[$winnerId]['stole'] += count($fromOthers);
-                        }
-
-                        $riichiStat[$winnerId]['stole'] += $item['from_table'];
-                    }
-                }
-            }
-        }
-
-        return $riichiStat;
-    }
-
-    /**
-     * @param $eventIdList
-     * @return array
-     * @throws \Exception
-     */
-    protected function _getRiichiStat1($eventIdList)
-    {
-        $games = SessionPrimitive::findByEventListAndStatus(
-            $this->_db,
-            $eventIdList,
-            SessionPrimitive::STATUS_FINISHED
-        );
-
-        $sessionIds = array_map(function (SessionPrimitive $el) {
-            return $el->getId();
-        }, $games);
-
-        $rounds = RoundPrimitive::findBySessionIds($this->_db, $sessionIds);
-        $players = EventModel::getPlayersOfGames($this->_db, $games);
-
-        $riichiStat = [];
-        foreach ($players as $playerId => $player) {
-            $riichiStat[$playerId] = [
-                'name'  => $player['display_name'],
-                'won'   => 0,
-                'lost'  => 0,
-            ];
-        }
-
-        foreach ($rounds as $round) {
-            $riichiIds = $round->getRiichiIds();
-
-            if (in_array($round->getOutcome(), ['ron', 'tsumo'])) {
-                $winnerId = $round->getWinnerId();
-
-                foreach ($riichiIds as $riichiPlayerId) {
-                    if ($riichiPlayerId == $winnerId) {
-                        $riichiStat[$riichiPlayerId]['won'] ++;
-                    } else {
-                        $riichiStat[$riichiPlayerId]['lost'] ++;
-                    }
-                }
-            }
-
-            if (in_array($round->getOutcome(), ['draw', 'abort'])) {
-                foreach ($riichiIds as $riichiPlayerId) {
-                    $riichiStat[$riichiPlayerId]['lost'] ++;
-                }
-            }
-
-            if ($round->getOutcome() == 'multiron') {
-                /** @var $round MultiRoundPrimitive */
-                $roundWinners = $round->getWinnerIds();
-
-                foreach ($riichiIds as $riichiPlayerId) {
-                    if (in_array($riichiPlayerId, $roundWinners)) {
-                        $riichiStat[$riichiPlayerId]['won'] ++;
-                    } else {
-                        $riichiStat[$riichiPlayerId]['lost'] ++;
-                    }
-                }
-            }
-        }
-
-        return $riichiStat;
-    }
-
-    /**
-     * @param string $winnerId
-     * @param int $fromTable
-     * @param [] $fromPlayers
-     * @param [][] $riichiBets
-     * @return array
-     */
-    protected function _addRiichiBets($winnerId, $fromTable, $fromPlayers, $riichiBets)
-    {
-        if (empty($riichiBets[$winnerId])) {
-            $riichiBets[$winnerId] = 0;
-        }
-
-        $fromOthers = array_filter($fromPlayers, function ($playerId) use ($winnerId) {
-            return $playerId != $winnerId;
-        });
-
-        $riichiBets[$winnerId] += $fromTable;
-        $riichiBets[$winnerId] += count($fromOthers);
-
-        return $riichiBets;
-    }
-
-
-    /**
-     * @param RoundPrimitive $round
-     * @param SessionState $sessionState
-     * @param [][] $payments
-     * @return array
-     * @throws \Exception
-     */
-    protected function _addLoserPayment($round, $sessionState, $payments)
-    {
-        if (!in_array($sessionState->getLastOutcome(), ['ron', 'multiron'])) {
-            return $payments;
-        }
-
-        $lastSessionState = $round->getLastSessionState();
-        $loserId = $round->getLoserId();
-        $loserHasRiichi = in_array($loserId, $round->getRiichiIds());
-
-        $lastScore = $lastSessionState->getScores()[$loserId];
-        $currentScore = $sessionState->getScores()[$loserId];
-
-        $payment = $lastScore - $currentScore;
-        if ($loserHasRiichi) {
-            $payment -= 1000;
-        }
-
-        if (empty($payments[$loserId])) {
-            $payments[$loserId] = [
-                'sum'   => 0,
-                'count' => 0
-            ];
-        }
-
-        $payments[$loserId]['sum'] += $payment;
-
-        if ($sessionState->getLastOutcome() !== 'multiron') {
-            $payments[$loserId]['count']++;
-        } else {
-            /** @var MultiRoundPrimitive $mRound */
-            $mRound = $round;
-            $rounds = $mRound->rounds();
-            $multiRonCount = $rounds[0]->getMultiRon();
-            $payments[$loserId]['count'] += $multiRonCount;
-        }
-
-        return $payments;
-    }
-
-    /**
-     * @param $sessionIds
-     * @return RoundPrimitive[][]
-     * @throws \Exception
-     */
-    protected function _getRounds($sessionIds)
-    {
-        $rounds = RoundPrimitive::findBySessionIds($this->_db, $sessionIds);
-
-        $result = [];
-        foreach ($rounds as $item) {
-            if (empty($result[$item->getSessionId()])) {
-                $result[$item->getSessionId()] = [];
-            }
-            $result[$item->getSessionId()] []= $item;
-        }
-
-        return $result;
-    }
-
-    /**
      * Find out currently playing tables state (for tournaments only)
      * @param integer $eventId
      * @param bool $includeAllRounds
